Changes to updatedatabase.pl script to work from 3.2.x to 21.5.x

+++ installer/data/mysql/updatedatabase.pl
@@ -1278,7 +1278,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
     $dbh->do("CREATE TABLE `user_permissions` (
                 `borrowernumber` int(11) NOT NULL DEFAULT 0,
                 `module_bit` int(11) NOT NULL DEFAULT 0,
-                `code` varchar(30) DEFAULT NULL,
+                `code` varchar(30) DEFAULT '' NOT NULL,
                 CONSTRAINT `user_permissions_ibfk_1` FOREIGN KEY (`borrowernumber`) REFERENCES `borrowers` (`borrowernumber`)
                     ON DELETE CASCADE ON UPDATE CASCADE,
                 CONSTRAINT `user_permissions_ibfk_2` FOREIGN KEY (`module_bit`, `code`)
@@ -2813,7 +2813,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = "3.01.00.072";
 if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
-    $dbh->do("INSERT INTO systempreferences (variable,value,explanation,options,type) VALUES('OpacPrivacy', '0', 'if ON, allows patrons to define their privacy rules (reading history)',NULL,'YesNo')");
+#    $dbh->do("INSERT INTO systempreferences (variable,value,explanation,options,type) VALUES('OpacPrivacy', '0', 'if ON, allows patrons to define their privacy rules (reading history)',NULL,'YesNo')");
     # create a new syspref for the 'Mr anonymous' patron
     $dbh->do("INSERT INTO systempreferences (variable,value,explanation,options,type) VALUES('AnonymousPatron', '0', \"Set the identifier (borrowernumber) of the 'Mister anonymous' patron. Used for Suggestion and reading history privacy\",NULL,'')");
     # fill AnonymousPatron with AnonymousSuggestion value (copy)
@@ -2827,7 +2827,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
     # 2nd, change the type to Choice
     $dbh->do("UPDATE systempreferences SET type='YesNo' WHERE variable='AnonSuggestions'");
         # borrower reading record privacy : 0 : forever, 1 : laws, 2 : don't keep at all
-    $dbh->do("ALTER TABLE `borrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
+#    $dbh->do("ALTER TABLE `borrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
     print "Upgrade to $DBversion done (add new syspref and column in borrowers)\n";
     SetVersion ($DBversion);
 }
@@ -3911,7 +3911,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = '3.03.00.006';
 if (C4::Context->preference("Version") < TransformToNum($DBversion) && $original_version < TransformToNum("3.02.01.002")) {
-    $dbh->do("ALTER TABLE deletedborrowers ADD `privacy` int(11) AFTER smsalertnumber;");
+#    $dbh->do("ALTER TABLE deletedborrowers ADD `privacy` int(11) AFTER smsalertnumber;");
     $dbh->do("ALTER TABLE deletedborrowers CHANGE `cardnumber` `cardnumber` varchar(16);");
     print "Upgrade to $DBversion done (Fix differences between borrowers and deletedborrowers)\n";
     SetVersion ($DBversion);
@@ -4006,8 +4006,8 @@ $DBversion = '3.03.00.016';
 if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
     # reimplement OpacPrivacy system preference
     $dbh->do("INSERT INTO systempreferences (variable,value,explanation,options,type) VALUES('OpacPrivacy', '0', 'if ON, allows patrons to define their privacy rules (reading history)',NULL,'YesNo')");
-    $dbh->do("ALTER TABLE `borrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
-    $dbh->do("ALTER TABLE `deletedborrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
+#    $dbh->do("ALTER TABLE `borrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
+#    $dbh->do("ALTER TABLE `deletedborrowers` ADD `privacy` INTEGER NOT NULL DEFAULT 1;");
     print "Upgrade to $DBversion done (OpacPrivacy reimplementation)\n";
     SetVersion($DBversion);
 };
@@ -4564,11 +4564,22 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = "3.07.00.001";
 if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
+    sanitize_zero_date('borrowers', 'dateexpiry');
+    sanitize_zero_date('borrowers', 'dateofbirth');
+    sanitize_zero_date('borrowers', 'dateenrolled');
     my $borrowers = $dbh->selectcol_arrayref( "SELECT borrowernumber from borrowers where debarred =1;", { Columns => [1] } );
+    $dbh->do( "UPDATE borrowers set debarred=NULL;" );
     $dbh->do("ALTER TABLE borrowers MODIFY debarred DATE DEFAULT NULL;");
     $dbh->do( "UPDATE borrowers set debarred='9999-12-31' where borrowernumber IN (" . join( ",", @$borrowers ) . ");" ) if ($borrowers and scalar(@$borrowers)>0);
     $dbh->do("ALTER TABLE borrowers ADD COLUMN debarredcomment VARCHAR(255) DEFAULT NULL AFTER debarred;");
+
+    sanitize_zero_date('deletedborrowers', 'dateexpiry');
+    sanitize_zero_date('deletedborrowers', 'dateofbirth');
+    sanitize_zero_date('deletedborrowers', 'dateenrolled');
+    $borrowers = $dbh->selectcol_arrayref( "SELECT borrowernumber from deletedborrowers where debarred =1;", { Columns => [1] } );
+    $dbh->do( "UPDATE deletedborrowers set debarred=NULL;" );
     $dbh->do("ALTER TABLE deletedborrowers MODIFY debarred DATE DEFAULT NULL;");
+    $dbh->do( "UPDATE deletedborrowers set debarred='9999-12-31' where borrowernumber IN (" . join( ",", @$borrowers ) . ");" ) if ($borrowers and scalar(@$borrowers)>0);
     $dbh->do("ALTER TABLE deletedborrowers ADD COLUMN debarredcomment VARCHAR(255) DEFAULT NULL AFTER debarred;");
     print "Upgrade done (Change borrowers.debarred into Date )\n";
     SetVersion($DBversion);
@@ -4611,6 +4622,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = "3.07.00.007";
 if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
+    sanitize_zero_date('items', 'dateaccessioned');
     $dbh->do("ALTER TABLE items MODIFY materials text;");
     print "Upgrade to $DBversion done alter items.material from varchar(10) to text \n";
     SetVersion($DBversion);
@@ -4739,8 +4751,8 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = "3.07.00.019";
 if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
-    $dbh->do(" UPDATE `systempreferences` SET  `value` =  'none', `options` =  'none|full|first|surname|firstandinitial|username', `explanation` =  'Choose how a commenter''s identity is presented alongside comments in the OPAC', `type` =  'Choice' WHERE  `systempreferences`.`variable` =  'ShowReviewer' AND `systempreferences`.`variable` = 0");
-    $dbh->do(" UPDATE `systempreferences` SET  `value` =  'full', `options` =  'none|full|first|surname|firstandinitial|username', `explanation` =  'Choose how a commenter''s identity is presented alongside comments in the OPAC', `type` =  'Choice' WHERE  `systempreferences`.`variable` =  'ShowReviewer' AND `systempreferences`.`variable` = 1");
+    $dbh->do(" UPDATE `systempreferences` SET  `value` =  'none', `options` =  'none|full|first|surname|firstandinitial|username', `explanation` =  'Choose how a commenter''s identity is presented alongside comments in the OPAC', `type` =  'Choice' WHERE  `systempreferences`.`variable` =  'ShowReviewer' AND `systempreferences`.`value` = 0");
+    $dbh->do(" UPDATE `systempreferences` SET  `value` =  'full', `options` =  'none|full|first|surname|firstandinitial|username', `explanation` =  'Choose how a commenter''s identity is presented alongside comments in the OPAC', `type` =  'Choice' WHERE  `systempreferences`.`variable` =  'ShowReviewer' AND `systempreferences`.`value` = 1");
     print "Upgrade to $DBversion done ( Adding additional options for the display of commenter's identity in the OPAC: Full name, first name, last name, first name and last name first initial, username, or no information)\n";
     SetVersion($DBversion);
 }
@@ -4790,7 +4802,7 @@ $DBversion = "3.07.00.023";
 if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
     $dbh->do("ALTER TABLE `message_transports` DROP FOREIGN KEY `message_transports_ibfk_3`");
     $dbh->do("ALTER TABLE `letter` DROP PRIMARY KEY");
-    $dbh->do("ALTER TABLE `letter` ADD `branchcode` varchar(10) default NULL AFTER `code`");
+    $dbh->do("ALTER TABLE `letter` ADD `branchcode` varchar(10) NOT NULL default '' AFTER `code`");
     $dbh->do("ALTER TABLE `letter` ADD PRIMARY KEY  (`module`,`code`, `branchcode`)");
     $dbh->do("ALTER TABLE `message_transports` ADD `branchcode` varchar(10) NOT NULL default ''");
     $dbh->do("ALTER TABLE `message_transports` ADD CONSTRAINT `message_transports_ibfk_3` FOREIGN KEY (`letter_module`, `letter_code`, `branchcode`) REFERENCES `letter` (`module`, `code`, `branchcode`) ON DELETE CASCADE ON UPDATE CASCADE");
@@ -5265,7 +5277,9 @@ if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
 
 $DBversion = "3.09.00.001";
 if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
-    $dbh->do("ALTER TABLE borrower_attribute_types MODIFY category_code VARCHAR( 1 ) NULL DEFAULT NULL");
+    $dbh->do("ALTER TABLE borrower_attribute_types DROP FOREIGN KEY `category_code_fk`");
+    $dbh->do("ALTER TABLE borrower_attribute_types MODIFY category_code VARCHAR( 10 ) NULL DEFAULT NULL");
+    $dbh->do("ALTER TABLE borrower_attribute_types ADD CONSTRAINT category_code_fk FOREIGN KEY (category_code) REFERENCES categories(categorycode)");
     print "Upgrade to $DBversion done. (Bug 8002: Update patron attribute types table to allow NULL category_code)\n";
     SetVersion($DBversion);
 }
@@ -5646,7 +5660,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
                  repeatable, mandatory, tab, authorised_value, value_builder,
                  seealso, isurl, hidden, linkid, kohafield, frameworkcode)
                 VALUES (?, ?, '9', '9 (RLIN)', '9 (RLIN)', 0, 0, ?, NULL, NULL,
-                    NULL, 0, 1, '', '', '')});
+                    NULL, 0, 1, 0, '', '')});
         my $sth3 = $dbh->prepare(q{UPDATE auth_subfield_structure SET
                                     frameworkcode = ? WHERE authtypecode = ? AND
                                     tagfield = ? AND tagsubfield = 'a'});
@@ -5760,7 +5774,7 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
               `tobranch` varchar(10) NOT NULL,
               `cost` decimal(6,2) NOT NULL,
               `disable_transfer` tinyint(1) NOT NULL DEFAULT 0,
-              CHECK ( `frombranch` <> `tobranch` ), -- a dud check, mysql does not support that
+#              CHECK ( `frombranch` <> `tobranch` ), -- a dud check, mysql does not support that
               PRIMARY KEY (`frombranch`, `tobranch`),
               CONSTRAINT `transport_cost_ibfk_1` FOREIGN KEY (`frombranch`) REFERENCES `branches` (`branchcode`) ON DELETE CASCADE ON UPDATE CASCADE,
               CONSTRAINT `transport_cost_ibfk_2` FOREIGN KEY (`tobranch`) REFERENCES `branches` (`branchcode`) ON DELETE CASCADE ON UPDATE CASCADE
@@ -5837,7 +5851,9 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
 $DBversion = "3.09.00.045";
 if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
+    $dbh->do("ALTER TABLE borrower_attribute_types DROP FOREIGN KEY `category_code_fk`");
     $dbh->do("ALTER TABLE borrower_attribute_types MODIFY category_code VARCHAR( 10 ) NULL DEFAULT NULL");
+    $dbh->do("ALTER TABLE borrower_attribute_types ADD CONSTRAINT category_code_fk FOREIGN KEY (category_code) REFERENCES categories(categorycode)");
     print "Upgrade to $DBversion done. (Bug 8002: Update patron attribute types table from varchar(1) to varchar(10) category_code)\nWarning to Koha System Administrators: If you use borrower attributes defined by borrower categories, you have to check your configuration. A bug may have removed your attribute links to borrower categories.\nPlease check, and fix it if necessary.";
     SetVersion($DBversion);
 }
@@ -5929,12 +5945,12 @@ if (C4::Context->preference("Version") < TransformToNum($DBversion)) {
 
     # Fill this new table with existing invoices
     my $sth = $dbh->prepare("
-        SELECT aqorders.booksellerinvoicenumber AS invoicenumber, aqbasket.booksellerid, aqorders.datereceived
+        SELECT aqorders.booksellerinvoicenumber AS invoicenumber, aqbasket.booksellerid, max(aqorders.datereceived)
         FROM aqorders
           LEFT JOIN aqbasket ON aqorders.basketno = aqbasket.basketno
         WHERE aqorders.booksellerinvoicenumber IS NOT NULL
           AND aqorders.booksellerinvoicenumber != ''
-        GROUP BY aqorders.booksellerinvoicenumber
+        GROUP BY aqorders.booksellerinvoicenumber,aqbasket.basketno
     ");
     $sth->execute;
     my $results = $sth->fetchall_arrayref({});
@@ -7611,7 +7627,7 @@ if ( CheckVersion($DBversion) ) {
         SELECT subscriptionid, startdate
         FROM subscription
         WHERE firstacquidate IS NULL
-          OR firstacquidate = '0000-00-00'
+          OR CAST(firstacquidate AS CHAR(10)) = '0000-00-00'
     |);
     $get_subscriptions_sth->execute;
     while ( my ($subscriptionid, $startdate) = $get_subscriptions_sth->fetchrow ) {
@@ -7826,6 +7842,7 @@ if ( CheckVersion($DBversion) ) {
 
 $DBversion = "3.15.00.002";
 if(CheckVersion($DBversion)) {
+    sanitize_zero_date('deleteditems','dateaccessioned');
     $dbh->do("ALTER TABLE deleteditems MODIFY materials text;");
     print "Upgrade to $DBversion done (Bug 11275: alter deleteditems.materials from varchar(10) to text)\n";
     SetVersion($DBversion);
@@ -8395,7 +8412,7 @@ if ( CheckVersion($DBversion) ) {
             message_transport_type="phone",
             name= ?
         WHERE code="OVERDUE_PHONE"
-    |, {}, $name);
+    |, {}, $name) if ( $name );
 
     print "Upgrade to $DBversion done (Bug 11867: Update letters *_PHONE)\n";
     SetVersion($DBversion);
@@ -9711,6 +9728,7 @@ if(CheckVersion($DBversion)) {
 
 $DBversion = "3.19.00.006";
 if ( CheckVersion($DBversion) ) {
+    sanitize_zero_date('categories', 'enrolmentperioddate');
     $dbh->do(q|SET foreign_key_checks = 0|);
     my $sth = $dbh->table_info( '','','','TABLE' );
     my ( $cat, $schema, $name, $type, $remarks );
@@ -10390,7 +10408,7 @@ $DBversion = "3.19.00.038";
 if ( CheckVersion($DBversion) ) {
     $dbh->do(q|
         ALTER TABLE virtualshelves
-        ADD COLUMN created_on DATETIME NOT NULL AFTER lastmodified
+        ADD COLUMN created_on DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP AFTER lastmodified
     |);
     # Set created_on = lastmodified
     # I would say it's better than 0000-00-00
@@ -10933,21 +10951,23 @@ if ( CheckVersion($DBversion) ) {
 
     my $AutoSelfCheckID = C4::Context->preference('AutoSelfCheckID');
 
-    $dbh->do(q|
-        UPDATE borrowers
-        SET flags=0
-        WHERE userid=?
-    |, undef, $AutoSelfCheckID);
+    if ( $AutoSelfCheckID ) {
+        $dbh->do(q|
+            UPDATE borrowers
+            SET flags=0
+            WHERE userid=?
+        |, undef, $AutoSelfCheckID);
 
-    $dbh->do(q|
-        DELETE FROM user_permissions
-        WHERE borrowernumber=(SELECT borrowernumber FROM borrowers WHERE userid=?)
-    |, undef, $AutoSelfCheckID);
+        $dbh->do(q|
+            DELETE FROM user_permissions
+            WHERE borrowernumber=(SELECT borrowernumber FROM borrowers WHERE userid=?)
+        |, undef, $AutoSelfCheckID);
 
-    $dbh->do(q|
-        INSERT INTO user_permissions(borrowernumber, module_bit, code)
-        SELECT borrowernumber, 1, 'self_checkout' FROM borrowers WHERE userid=?
-    |, undef, $AutoSelfCheckID);
+        $dbh->do(q|
+            INSERT INTO user_permissions(borrowernumber, module_bit, code)
+            SELECT borrowernumber, 1, 'self_checkout' FROM borrowers WHERE userid=?
+        |, undef, $AutoSelfCheckID);
+    }
     print "Upgrade to $DBversion done (Bug 14298: AutoSelfCheckID user should only be able to access SCO)\n";
     SetVersion($DBversion);
 }
@@ -12102,7 +12122,7 @@ if ( C4::Context->preference("Version") < TransformToNum($DBversion) ) {
     my ($print_error) = $dbh->{PrintError};
     $dbh->{RaiseError} = 0;
     $dbh->{PrintError} = 0;
-    $dbh->do("ALTER TABLE overduerules_transport_types ADD COLUMN letternumber INT(1) NOT NULL DEFAULT 1 AFTER id");
+    eval { $dbh->do("ALTER TABLE overduerules_transport_types ADD COLUMN letternumber INT(1) NOT NULL DEFAULT 1 AFTER id"); };
     $dbh->{PrintError} = $print_error;
 
     print "Upgrade to $DBversion done (Bug 16007: Make sure overduerules_transport_types.letternumber exists)\n";
@@ -12616,9 +12636,9 @@ if ( CheckVersion($DBversion) ) {
     $dbh->do(q{
         UPDATE letter SET branchcode='' WHERE branchcode IS NULL;
     });
-    $dbh->do(q{
-        ALTER TABLE letter MODIFY COLUMN branchcode varchar(10) NOT NULL DEFAULT ''
-    });
+#    $dbh->do(q{
+#        ALTER TABLE letter MODIFY COLUMN branchcode varchar(10) NOT NULL DEFAULT ''
+#    });
     $dbh->do(q{
         ALTER TABLE permissions MODIFY COLUMN code varchar(64) NOT NULL DEFAULT '';
     });
